+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+### `Project Import` enhancements
+- Loading backoffice sources provided by SAP on project import [#1096](https://github.com/epam/sap-commerce-intellij-idea-plugin/issues/1096)
+
 ## [2024.1.1]
 
 ### `Project Import` enhancements